test_float64

+++ snpe_reputation_systems/tests/test_simulator_class.py
@@ -7,7 +7,7 @@ import pytest
 from hypothesis import given, settings
 from hypothesis import strategies as st
 from hypothesis.extra.numpy import arrays
-from numpy import float
+from numpy import float64
 
 from ..snpe_reputation_systems.simulations.simulator_class import (
     BaseSimulator,
@@ -30,7 +30,7 @@ from ..snpe_reputation_systems.simulations.simulator_class import (
 #    return BaseSimulator(params)
 
 
-@given(arrays(float, 5), arrays(float, 6), st.none())
+@given(arrays(float64, 5), arrays(float64, 6), st.none())
 def test_convolve_prior_with_existing_reviews(arr1, arr2, none_value):
     # BaseSimulator instance
     params = {
